Working on it added validation image etc
Is unfinished. I added some validations but need more. If Arduino is not connected should not have errors but put messages n/a I need to do it. Also I need to recize image and modify standard libraries that had mistakes over all when microcontroller is not connected by any reason throwing a ioerror. Also did have some macros or others program called macros to define functions and pin analogic assignments. After that I need to do a window menu to open a file and append info and plot with matplotlib old versions 32 bit fails to download from pip (1.5, 1.2 for example) i tried to correct the download manager and manual install the source libraries need but didn't work to failing trying to find visual c and some libraries and matching some versions. So I will do it but I will not have chance to compile and test this part.