Serial revised and custombox  backup f/security
They are erasing the files. Yesterday did this 3 times. So despite is not important I and distributing it again.
I revised the serial and now have even PARITY stop bit, and I changed the handling. I think that now is correct.
I revised the initialization of the dictionary arrays since we're wrong. I initialize them out of the for statement and then set them to cero and make the needed keys.
And changed the customs box class to make message boxes with 4 buttons.
Now is starting to see buttons and messages.