repairing serial and files to test with board
Serial had one problem when i test it about recursive way with fail. Was necessary add RETURN eadh time of the function call back itself to finish each failed step. Also was necesary add one more with an if and i changed one while for one to avoid stop there because of are inside of a Try and try again if no serial exception happens. Now I working in receiveData i have to change arduino program and all serial functions in the same way. Also I changed ports in meno from a message box from a RadioButton direct on the menu, since use StringVar to change it to String to can set the port I had to add the variable get change In to function that loop communication because RadioButton dont have command function to change StringVar() to String necessary to set the port. ALL TEST ARE WITH PRINTS AND I DIDNT ERASE THEM TO SHOW HOW I AM TESTING.