Phyton fails checking port in IDLE
Now the phyton is failing recognizing the port in the IDLE but not in the terminal. Can be seen the video

https://youtube.com/shorts/wpxnRmaYzxY?si=moxZ966QfS2vvR-s

I waste plenty time in this with this vulnerability that makes it untrustable.

Now computer system are highly corrupted and whose handle don't respect the rules of the codes and easily with keys or troy horses in the code can do serious damage, rob banks account, take money from dead persons or kill them before of that have on click university title and plenty more frauds.

This is a small example that show the disadvantage about some decades before of use a paper file, a record book and do a line for the bank or assist to a concrete place to receive a course. THIS DOESN'T DEMERITS THOSE THAT SHOW PRACTICAL EXAMPLES OF SELF LEARNING AND DESERVE RESPECT.